+++ src/deprecated/HYPOTv2.ts
@@ -157,7 +157,7 @@ export class HYPOTv2
     for (let i = inwardA.length; i--;) {
       const c = inwardA[i];
       if (c.from === c.to) {
-        neuron.creature.disconnect(c.from, c.to);
+        neuron.creature.disconnect(c.from, c.to, c.type);
       }
     }